+++ CHANGELOG.md
@@ -5,6 +5,6 @@ CHANGELOG
 * Update Terraform bridge to be based on v1.0.0 of the Terraform Plugin SDK
 * Add support to specify a custom package name for NodeJS package
 * Add ability to pass the TF provider version that the pulumi provider was generated against
+* Allow a schema variable to be overridden as being `Computed`
 
 ---
-